requirements.txt
Django==1.9
django-filter==0.11.0
djangorestframework==3.3.1
httpie==0.9.2
Markdown==2.6.5
psycopg2==2.6.1
Pygments==2.0.2
requests==2.8.1
wheel==0.24.0
